backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1
google-generativeai==0.7.1
python-pptx==0.6.22
python-docx==1.1.0
PyPDF2==3.0.1
requests==2.31.0
aiofiles==23.2.1
pillow==9.5.0
pydantic==2.6.1 